app/requirements.txt
Flask
flask-socketio
eventlet
mysql-connector-python
